requirements.txt
PyQt6>=6.4.0
PyQt6-Qt6>=6.4.0
pytesseract>=0.3.10
Pillow>=10.0.0
argostranslate>=1.9.0
googletrans==4.0.0rc1
pynput>=1.7.6
setproctitle>=1.3.2
requests>=2.31.0
